Switch to new fabric search mechanism (leave the old one in place). Switch vendors really don't like to support GET ALL NEXT, so we use a different, larger memory footprint, method.
Split 2300/2312 support so we distinguish between the two.

Correctly identify that SCC Luns are 16384 luns for us, not 65536.

Turn on Reduced Interrupt Operation for LVD SCSI cards, which rocks.

Do 'mailbox continuations' in some cases- this batches repeated mailbox
commands so we don't have to wake the invokee until we're done with the
N thousand or so mailbox commands we needed to do.

Put in firmware dump code (optioned out for now- but there if needed to
capture issues for QLogic).

Fix a boatload of bugs- like handling cases of dropped frames that show
up as 'bogus' residuals (i.e., we get a a DATA UNDERRUN as reported by
the f/w, but there is no RESPONSE UNDERRUN in the FCP RSPNS IU, or the
residual is bogus- this led to a lot of silent data corruption cases).