Plug obvious memory leaks.
The comparison function was previously changed by me to use strcmp,
which is probably wrong since a buffer can theoretically have NULs
in it.  This fixes it.  (In my defense, the original code converted
the buffer text to char * values and compared until there was either
a difference or until a NUL was reached, so my previous change was
no worse.)

OK joris@