set the blocksize to the size of the read/write buffers.
the blocksize matters for the poll() cases because we pass the poll
when reading from the device when there is any data in the record buffer.
since data is transfered into the buffer in blocksize chunks, if there
was no and now is some data in the buffer, there is probably about
blocksize bytes of data in the buffer.  if in non-blocking mode, which
is common when using poll(), and we try to read() more bytes than are in
the record buffer, we will receive an EAGAIN.  if we read() in
blocksize chunks, we will never try to read more than is available.

a more eloquent version of the above description should be in audio(4).